feat(gangan) -歌单详情搞掂。

+++ common/src/main/java/com/example/common/bean/searchbean/SearchMusicBean.kt
@@ -49,7 +49,7 @@ data class Quality(
     val fid: Int,
     val size: Int,
     val sr: Int,
-    val vd: Int
+    val vd: Double
 )
